requirements.txt
torch>=2.0.0
transformers>=4.35.0
tqdm
pillow
numpy
scikit-learn
opencv-python
torchvision
